requirements.txt
Django==2.2.4
djangorestframework==3.10.2
channels==2.2.0
django-split-settings==0.3.0
psycopg2-binary==2.8.3
django-cors-headers==3.1.0
